DAI: fightrelative action adds to queue rather than executing immediately
Let's assume we have a building to build, and then the next task will be a
fightrelative (assistqueue). TaskQueueBehaviour tries to be one step ahead so
when a task finishes, the next one is already there (queued, at least with
build orders). So when the unit starts building, TaskQueueBehaviour looks at
the next task and executes it, overriding the recently started build job. This
casued buildings to be left alone, the unit went away and helped constructing
in a factory, and when the factory finished, it returned to the incomplete
build and "helped" finish it. If the factory unit takes too long, to building
would have disintegrated by then.

Change fightrelative task to only queue itself, so the unit will finish the
building first. Code analysis showed that fightrelative is only used for
construction units, for assisting others or patroling, so I on't expect any
regressions elsewhere.