DAI: only use wind if it's worth it longterm
Take average wind into account when deciding  what kind of energy source to
build. Use wind only if average is above 8. Mathematically >= 7 would be enough
but I'm using a little higher threshold to compensate for is unpredictability.
It would still be useful to mix with some solars not just based on current wind
speed.