Improve squad assignments
Units are only assigned to a squad if they are close to each other. This will
rationalize their movement (no unit has to move through the map just because
the first squad was there), and also help minimize the regrouping effect.
Regrouping is great to slow down fast units, and make them wait for the slower
ones, but if a unit in the base is assigned to a squad that is near the enemy,
regrouping has the effect of withdrawing forces from enemies.

Also increase the group size to allow more units to attack the same target.